+++ JavaAbstractMachine/src/at/mhofer/jam/data/constantpool/CpNameAndTypeInfo.java
@@ -10,7 +10,7 @@ public class CpNameAndTypeInfo implements CpInfo
 	 * method name <init> (§2.9) or a valid unqualified name denoting a field or
 	 * method (§4.2.2).
 	 */
-	private short nameIndex;
+	private int nameIndex;
 
 	/**
 	 * The value of the descriptor_index item must be a valid index into the
@@ -18,9 +18,9 @@ public class CpNameAndTypeInfo implements CpInfo
 	 * CONSTANT_Utf8_info structure (§4.4.7) representing a valid field
 	 * descriptor or method descriptor (§4.3.2, §4.3.3).
 	 */
-	private short descriptorIndex;
+	private int descriptorIndex;
 
-	public CpNameAndTypeInfo(short nameIndex, short descriptorIndex)
+	public CpNameAndTypeInfo(int nameIndex, int descriptorIndex)
 	{
 		this.nameIndex = nameIndex;
 		this.descriptorIndex = descriptorIndex;
@@ -32,12 +32,12 @@ public class CpNameAndTypeInfo implements CpInfo
 		return CpTag.NAME_AND_TYPE;
 	}
 
-	public short getNameIndex()
+	public int getNameIndex()
 	{
 		return nameIndex;
 	}
 
-	public short getDescriptorIndex()
+	public int getDescriptorIndex()
 	{
 		return descriptorIndex;
 	}
